requirements-dev.txt
# Test dependencies for Otto (API + ASR/STT tests)
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0
# Optional: for WER (Word Error Rate) in ASR evaluation
jiwer>=3.0.0
